autocorrect-server/requirements.txt
python-dotenv==0.15.0
regex==2020.11.13
websockets==8.1
